requirements.txt
pyyaml
fastapi
slack-bolt
google-cloud-storage
openai
scikit-learn
mangum==0.17.0